# Conflicts:
#	shared/public/SimpleLayerInterface.h
#	shared/public/Tiled2dMapSourceImpl.h
#	shared/public/Tiled2dMapVectorLayer.h
#	shared/public/Tiled2dMapVectorLayerParserHelper.h
#	shared/src/map/layers/tiled/vector/Tiled2dMapVectorLayer.cpp

+++ bridging/ios/MCTiled2dMapRasterLayerInterface.h
@@ -50,9 +50,9 @@
 
 - (void)removeCallbackHandler;
 
-- (void)setAlpha:(double)alpha;
+- (void)setAlpha:(float)alpha;
 
-- (double)getAlpha;
+- (float)getAlpha;
 
 - (void)setMinZoomLevelIdentifier:(nullable NSNumber *)value;
