+++ shared/public/Tiled2dMapRasterLayerInterface.h
@@ -46,9 +46,9 @@ public:
 
     virtual void removeCallbackHandler() = 0;
 
-    virtual void setAlpha(double alpha) = 0;
+    virtual void setAlpha(float alpha) = 0;
 
-    virtual double getAlpha() = 0;
+    virtual float getAlpha() = 0;
 
     virtual void setMinZoomLevelIdentifier(std::optional<int32_t> value) = 0;
